Changed %lu formatters to %llu

+++ cluster_dynamics.cpp
@@ -63,10 +63,10 @@ double i_clusters_delta(uint64_t in)
     double n = iemission_vabsorption_n(in);
     double nm1 = in == 1 ? 0. : iemission_vabsorption_nm1(in - 1);
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "i_defect_production(%lu)%.*s%g +\n", in, 12, TABS, g);
-    fprintf(stderr, "iemission_vabsorption_np1(%lu) * interstitials[%lu]%.*s%g * %g -\n", in + 1, in + 1, 9, TABS, np1, interstitials[in + 1]);
-    fprintf(stderr, "iemission_vabsorption_n(%lu) * interstitials[%lu]%.*s%g * %g", in, in, 9, TABS, n, interstitials[in]);
-    if (in > 1) fprintf(stderr, " +\niemission_vabsorption_nm1(%lu) * interstitials[%lu]%.*s%g * %g",  in - 1, in - 1, 9, TABS, nm1, interstitials[in - 1]);
+    fprintf(stderr, "i_defect_production(%llu)%.*s%g +\n", (unsigned long long) in, 12, TABS, g);
+    fprintf(stderr, "iemission_vabsorption_np1(%llu) * interstitials[%llu]%.*s%g * %g -\n", (unsigned long long) in + 1, (unsigned long long) in + 1, 9, TABS, np1, interstitials[in + 1]);
+    fprintf(stderr, "iemission_vabsorption_n(%llu) * interstitials[%llu]%.*s%g * %g", (unsigned long long) in, (unsigned long long) in, 9, TABS, n, interstitials[in]);
+    if (in > 1) fprintf(stderr, " +\niemission_vabsorption_nm1(%llu) * interstitials[%llu]%.*s%g * %g",  (unsigned long long) in - 1, (unsigned long long) in - 1, 9, TABS, nm1, interstitials[in - 1]);
     fprintf(stderr, "\n");
     return print_return(g + np1 * interstitials[in + 1] - n * interstitials[in] + (in == 1 ? 0. : nm1 * interstitials[in - 1]));
     #else
@@ -97,10 +97,13 @@ double v_clusters_delta(uint64_t vn)
     double n = vemission_iabsorption_n(vn);
     double nm1 = vn == 1 ? 0. : vemission_iabsorption_nm1(vn - 1);
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "v_defect_production(%lu)%.*s%g +\n", vn, 12, TABS, g);
-    fprintf(stderr, "vemission_iabsorption_np1(%lu) * vacancies[%lu]%.*s%g * %g -\n",  vn + 1, vn + 1, 9, TABS, np1, vacancies[vn + 1]);
-    fprintf(stderr, "vemission_iabsorption_n(%lu) * vacancies[%lu]%.*s%g * %g", vn, vn, 9, TABS, n, vacancies[vn]);
-    if (vn > 1) fprintf(stderr, " +\nvemission_iabsorption_nm1(%lu) * vacancies[%lu]%.*s%g * %g", vn - 1, vn - 1, 9, TABS, nm1, vacancies[vn - 1]);
+    fprintf(stderr, "v_defect_production(%llu)%.*s%g +\n", (unsigned long long) vn, 12, TABS, g);
+    fprintf(stderr, "vemission_iabsorption_np1(%llu) * vacancies[%llu]%.*s%g * %g -\n",  (unsigned long long) vn + 1, (unsigned long long) vn + 1, 9, TABS, np1, vacancies[vn + 1]);
+    fprintf(stderr, "vemission_iabsorption_n(%llu) * vacancies[%llu]%.*s%g * %g", (unsigned long long) vn, (unsigned long long) vn, 9, TABS, n, vacancies[vn]);
+    if (vn > 1) 
+        fprintf(stderr, " +\nvemission_iabsorption_nm1(%llu) * vacancies[%llu]%.*s%g * %g", 
+        (unsigned long long) vn - 1, (unsigned long long) vn - 1, 9, 
+        TABS, nm1, vacancies[vn - 1]);
     fprintf(stderr, "\n");
     return print_return(g + np1 * vacancies[vn + 1] - n * vacancies[vn] + (vn == 1 ? 0. : nm1 * vacancies[vn - 1]));
     #else
@@ -133,9 +136,9 @@ double iemission_vabsorption_np1(uint64_t np1)
     double vc1 = vacancies[1];
     double iie = ii_emission(np1);
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "iv_absorption(%lu)%.*s%g *\n", np1, 12, TABS, iva);
-    fprintf(stderr, "v_clusters1(%lu)%.*s%g +\n", np1, 13, TABS, vc1);
-    fprintf(stderr, "ii_emission(%lu)%.*s%g\n", np1, 13, TABS, iie);
+    fprintf(stderr, "iv_absorption(%llu)%.*s%g *\n", (unsigned long long) np1, 12, TABS, iva);
+    fprintf(stderr, "v_clusters1(%llu)%.*s%g +\n", (unsigned long long) np1, 13, TABS, vc1);
+    fprintf(stderr, "ii_emission(%llu)%.*s%g\n", (unsigned long long) np1, 13, TABS, iie);
     return print_return(iva * vc1 + iie);
     #else
 
@@ -163,9 +166,9 @@ double vemission_iabsorption_np1(uint64_t np1)
     double ic1 = interstitials[1];
     double vve = vv_emission(np1);
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "vi_absorption(%lu)%.*s%g *\n", np1, 12, TABS, via);
-    fprintf(stderr, "i_clusters1(%lu)%.*s%g +\n", np1, 13, TABS, ic1);
-    fprintf(stderr, "vv_emission(%lu)%.*s%g\n", np1, 13, TABS, vve);
+    fprintf(stderr, "vi_absorption(%llu)%.*s%g *\n", (unsigned long long) np1, 12, TABS, via);
+    fprintf(stderr, "i_clusters1(%llu)%.*s%g +\n", (unsigned long long) np1, 13, TABS, ic1);
+    fprintf(stderr, "vv_emission(%llu)%.*s%g\n", (unsigned long long) np1, 13, TABS, vve);
     return print_return(via * ic1 + vve);
     #else
     return 
@@ -195,9 +198,9 @@ double iemission_vabsorption_n(uint64_t n)
     double ic1 = interstitials[1];
     double iie = ii_emission(n);
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "iv_absorption(%lu) * v_clusters1(%lu)%.*s%g * %g +\n", n, n, 10, TABS, iva, vc1);
-    fprintf(stderr, "ii_absorption(%lu) * i_clusters1(%lu)%.*s%g * %g +\n", n, n, 10, TABS, iia, ic1);
-    fprintf(stderr, "ii_emission(%lu)%.*s%g\n", n, 13, TABS, iie);
+    fprintf(stderr, "iv_absorption(%llu) * v_clusters1(%llu)%.*s%g * %g +\n", (unsigned long long) n, (unsigned long long) n, 10, TABS, iva, vc1);
+    fprintf(stderr, "ii_absorption(%llu) * i_clusters1(%llu)%.*s%g * %g +\n", (unsigned long long) n, (unsigned long long) n, 10, TABS, iia, ic1);
+    fprintf(stderr, "ii_emission(%llu)%.*s%g\n", (unsigned long long) n, 13, TABS, iie);
     return print_return(iva * vc1 + iia * ic1 + iie);
     #else
     return 0;
@@ -227,9 +230,9 @@ double vemission_iabsorption_n(uint64_t n)
     double vc1 = vacancies[1];
     double vve = vv_emission(n);
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "vi_absorption(%lu) * i_clusters1(%lu)%.*s%g * %g +\n", n, n, 10, TABS, via, ic1);
-    fprintf(stderr, "vv_absorption(%lu) * v_clusters1(%lu)%.*s%g * %g +\n", n, n, 10, TABS, vva, vc1);
-    fprintf(stderr, "vv_emission(%lu)%.*s%g\n", n, 13, TABS, vve);
+    fprintf(stderr, "vi_absorption(%llu) * i_clusters1(%llu)%.*s%g * %g +\n", (unsigned long long) n, (unsigned long long) n, 10, TABS, via, ic1);
+    fprintf(stderr, "vv_absorption(%llu) * v_clusters1(%llu)%.*s%g * %g +\n", (unsigned long long) n, (unsigned long long) n, 10, TABS, vva, vc1);
+    fprintf(stderr, "vv_emission(%llu)%.*s%g\n", (unsigned long long) n, 13, TABS, vve);
     return print_return(via * ic1 + vva * vc1 + vve);
     #else
     return 
@@ -255,8 +258,8 @@ double iemission_vabsorption_nm1(uint64_t nm1)
     double iia = ii_absorption(nm1);
     double ic1 = interstitials[1];
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "ii_absorption(%lu)%.*s%g *\n", nm1, 12, TABS, iia);
-    fprintf(stderr, "i_clusters1(%lu)%.*s%g\n", nm1, 13, TABS, ic1);
+    fprintf(stderr, "ii_absorption(%llu)%.*s%g *\n", (unsigned long long) nm1, 12, TABS, iia);
+    fprintf(stderr, "i_clusters1(%llu)%.*s%g\n", (unsigned long long) nm1, 13, TABS, ic1);
     return print_return(iia * ic1);
     #else
     return
@@ -280,8 +283,8 @@ double vemission_iabsorption_nm1(uint64_t nm1)
     double vva = vv_absorption(nm1);
     double vc1 = vacancies[1];
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "vv_absorption(%lu)%.*s%g *\n", nm1, 12, TABS, vva);
-    fprintf(stderr, "v_clusters1(%lu)%.*s%g\n", nm1, 13, TABS, vc1);
+    fprintf(stderr, "vv_absorption(%llu)%.*s%g *\n", (unsigned long long) nm1, 12, TABS, vva);
+    fprintf(stderr, "v_clusters1(%llu)%.*s%g\n", (unsigned long long) nm1, 13, TABS, vc1);
     return print_return(vva * vc1);
     #else
     return
@@ -643,11 +646,11 @@ double ii_emission(uint64_t n)
             (BOLTZMANN_EV_KELVIN * reactor.temperature)
         );
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "2 * M_PI * %lu%.*s%g *\n", n, 13, TABS, pi2n);
-    fprintf(stderr, "i_bias_factor(%lu)%.*s%g *\n", n, 12, TABS, ibf);
+    fprintf(stderr, "2 * M_PI * %llu%.*s%g *\n", (unsigned long long) n, 13, TABS, pi2n);
+    fprintf(stderr, "i_bias_factor(%llu)%.*s%g *\n", (unsigned long long) n, 12, TABS, ibf);
     fprintf(stderr, "material.i_diffusion / material.atomic_volume%.*s%g *\n", 10, TABS, id);
-    fprintf(stderr, "-i_binding_energy(%lu)%.*s%g *\n", n, 12, TABS, -ibe);
-    fprintf(stderr, "exp(-i_binding_energy(%lu) / (k * reactor.temperature))%.*s%g *\n", n, 8, TABS, evexp);
+    fprintf(stderr, "-i_binding_energy(%llu)%.*s%g *\n", (unsigned long long) n, 12, TABS, -ibe);
+    fprintf(stderr, "exp(-i_binding_energy(%llu) / (k * reactor.temperature))%.*s%g *\n", (unsigned long long) n, 8, TABS, evexp);
     return print_return(pi2n * ibf * id * evexp);
     #else
 
@@ -704,11 +707,11 @@ double vv_emission(uint64_t n)
             (BOLTZMANN_EV_KELVIN * reactor.temperature)
         );
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
-    fprintf(stderr, "2 * M_PI * %lu%.*s%g *\n", n, 13, TABS, pi2n);
-    fprintf(stderr, "v_bias_factor(%lu)%.*s%g *\n", n, 12, TABS, vbf);
+    fprintf(stderr, "2 * M_PI * %llu%.*s%g *\n", (unsigned long long) n, 13, TABS, pi2n);
+    fprintf(stderr, "v_bias_factor(%llu)%.*s%g *\n", (unsigned long long) n, 12, TABS, vbf);
     fprintf(stderr, "material.v_diffusion%.*s%g *\n", 12, TABS, vd);
-    fprintf(stderr, "-v_binding_energy(%lu)%.*s%g *\n", n, 12, TABS, -vbe);
-    fprintf(stderr, "exp(-v_binding_energy(%lu) / (k * reactor.temperature))%.*s%g *\n", n, 8, TABS, evexp);
+    fprintf(stderr, "-v_binding_energy(%llu)%.*s%g *\n", (unsigned long long) n, 12, TABS, -vbe);
+    fprintf(stderr, "exp(-v_binding_energy(%llu) / (k * reactor.temperature))%.*s%g *\n", (unsigned long long) n, 8, TABS, evexp);
     return print_return(pi2n * vbf * vd * evexp);
     #else
     return 
@@ -775,7 +778,7 @@ double i_bias_factor(uint64_t in)
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
     fprintf(stderr, "material.i_dislocation_bias%.*s%g +\n", 11, TABS, idb);
     fprintf(stderr, "sqrt(burgers_vector / (8 * M_PI * material.lattice_param)) * material.i_loop_bias - material.i_dislocation_bias%.*s%g *\n", 1, TABS, llp);
-    fprintf(stderr, "1 / pow(%lu, material.i_dislocation_bias_param / 2)%.*s%g\n", in, 8, TABS, nexp);
+    fprintf(stderr, "1 / pow(%llu, material.i_dislocation_bias_param / 2)%.*s%g\n", (unsigned long long) in, 8, TABS, nexp);
     return print_return(idb + llp * nexp);
     #else
     return 
@@ -825,7 +828,7 @@ double v_bias_factor(uint64_t vn)
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
     fprintf(stderr, "material.v_dislocation_bias%.*s%g +\n", 11, TABS, vdb);
     fprintf(stderr, "sqrt(burgers_vector / (8 * M_PI * material.lattice_param)) * material.v_loop_bias - material.v_dislocation_bias%.*s%g *\n", 1, TABS, llp);
-    fprintf(stderr, "1 / pow(%lu, material.v_dislocation_bias_param / 2)%.*s%g\n", vn, 8, TABS, nexp);
+    fprintf(stderr, "1 / pow(%llu, material.v_dislocation_bias_param / 2)%.*s%g\n", (unsigned long long) vn, 8, TABS, nexp);
     return print_return(vdb + llp * nexp);
     #else
     return 
@@ -863,7 +866,7 @@ double i_binding_energy(uint64_t in)
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
     fprintf(stderr, "material.i_formation%.*s%g +\n", 12, TABS, ifo);
     fprintf(stderr, "(material.i_binding - material.i_formation) / (pow(2., .8) - 1)%.*s%g *\n", 6, TABS, factor);
-    fprintf(stderr, "(pow(%lu, .8) - pow(%lu, .8))%.*s%g *\n", in, in - 1, 11, TABS, npow);
+    fprintf(stderr, "(pow(%llu, .8) - pow(%llu, .8))%.*s%g *\n", (unsigned long long) in, (unsigned long long) in - 1, 11, TABS, npow);
     return print_return(ifo + factor * npow);
     #else
     return
@@ -885,7 +888,7 @@ double v_binding_energy(uint64_t vn)
     fprintf(stderr, "%s\t%s:%d\n", __FUNCTION__, __FILE__, __LINE__);
     fprintf(stderr, "material.v_formation%.*s%g +\n", 12, TABS, vfo);
     fprintf(stderr, "(material.v_binding - material.v_formation) / (pow(2., .8) - 1)%.*s%g *\n", 6, TABS, factor);
-    fprintf(stderr, "(pow(%lu, .8) - pow(%lu, .8))%.*s%g *\n", vn, vn - 1, 11, TABS, npow);
+    fprintf(stderr, "(pow(%llu, .8) - pow(%llu, .8))%.*s%g *\n", (unsigned long long) vn, (unsigned long long) vn - 1, 11, TABS, npow);
     return print_return(vfo + factor * npow);
     #else
     return
