+++ main.cpp
@@ -49,7 +49,7 @@ void print_start_message()
     fprintf(stderr, "\nSimulation Started: ");
     fprintf(stderr, "delta_time: %g, ", delta_time);
     fprintf(stderr, "simulation_time: %g, ", simulation_time);
-    fprintf(stderr, "concentration_boundary: %lu\n\n", concentration_boundary);
+    fprintf(stderr, "concentration_boundary: %llu\n\n", (unsigned long long) concentration_boundary);
 }
 
 int main(int argc, char* argv[])
@@ -81,7 +81,7 @@ int main(int argc, char* argv[])
         for (uint64_t n = 2; n < concentration_boundary && valid_sim; ++n)
         {
             #if VPRINT
-            fprintf(stdout, "\n------------------------------------------------------------------------------- t = %g\tn = %lu\n", t, n);
+            fprintf(stdout, "\n------------------------------------------------------------------------------- t = %g\tn = %llu\n", t, (unsigned long long) n);
             #endif
 
             interstitials_temp[n] += i_clusters_delta(n) * delta_time;
@@ -94,9 +94,9 @@ int main(int argc, char* argv[])
 
             #if CSV
                 #ifdef N
-                fprintf(stdout, "%g,%lu,%g,%g\n", t, (uint64_t)N, interstitials_temp[N], vacancies_temp[N]);
+                fprintf(stdout, "%g,%llu,%g,%g\n", t, (uint64_t)N, interstitials_temp[N], vacancies_temp[N]);
                 #else
-                fprintf(stdout, "%g,%lu,%g,%g\n", t, n, interstitials_temp[n], vacancies_temp[n]);
+                fprintf(stdout, "%g,%llu,%g,%g\n", t, n, interstitials_temp[n], vacancies_temp[n]);
                 #endif
             #endif
         }
@@ -117,7 +117,7 @@ int main(int argc, char* argv[])
     fprintf(stdout, "\nCluster Size\t\t-\t\tInterstitials\t\t-\t\tVacancies\n\n");
     for (uint64_t n = 1; n < concentration_boundary; ++n)
     {
-        fprintf(stdout, "%lu\t\t\t\t\t%13g\t\t\t  %15g\n\n", n, interstitials[n], vacancies[n]);
+        fprintf(stdout, "%llu\t\t\t\t\t%13g\t\t\t  %15g\n\n", (unsigned long long) n, interstitials[n], vacancies[n]);
     }
     #endif
     // --------------------------------------------------------------------------------------------
